+++ src/components/Home.vue
@@ -13,9 +13,6 @@
 </template>
 
 <script>
-import firebase from 'firebase/app'
-import 'firebase/auth'
-
 export default {
   name: 'home',
   data () {
@@ -35,27 +32,6 @@ export default {
   methods: {
     signOut () {
       this.$store.dispatch('signOut')
-    }
-  },
-  mounted () {     
-    // ローカルストレージに保存してあるメールアドレスを取得
-    const email = window.localStorage.getItem("emailForSignIn");
-    console.log('emailの中身')
-    console.log(email)
-    if(firebase.auth().isSignInWithEmailLink(window.location.href)) {
-        firebase.auth().signInWithEmailLink(email, window.location.href).then((result) => {
-            // ログイン完了
-            // ローカルに保存したメールアドレスを削除
-            window.localStorage.removeItem("emailForSignIn");
-        }).catch((error) => {
-            // ローカルに保存したメールアドレスを削除
-            window.localStorage.removeItem("emailForSignIn");
-            console.log(`signInWithEmailLinkでエラー発生：${error}`)
-        });
-    } else {
-        // ローカルに保存したメールアドレスを削除
-        window.localStorage.removeItem("emailForSignIn");
-        console.log('リンクがメールリンクによるログインではない普通のログイン？')
     }
   }
 }
